# Add Numba to rolling.apply