--- a/pandas/core/window/aggregators/methods.py
+++ b/pandas/core/window/aggregators/methods.py
@@ -1,3 +1,6 @@
+from typing import Callable
+
+import numba
 import numpy as np
 
 
@@ -32,4 +35,24 @@ def rolling_mean(
         if count and count >= minimum_periods:
             val = total / count
         result[i] = val
+    return result
+
+
+@numba.njit(nogil=True)
+def rolling_apply(
+    values: np.ndarray,
+    begin: np.ndarray,
+    end: np.ndarray,
+    minimum_periods: int,
+    numba_func: Callable,
+    args,
+) -> np.ndarray:
+    result = np.empty(len(begin))
+    for i, (start, stop) in enumerate(zip(begin, end)):
+        window = values[start:stop]
+        count_nan = np.sum(np.isnan(window))
+        if len(window) - count_nan >= minimum_periods:
+            result[i] = numba_func(window, *args)
+        else:
+            result[i] = np.nan
     return result