@mroeschke 2019-09-18
Here's the performance comparison so far. Unfortunately cython is still faster for the ndarray case.
```
# This branch
In [1]: s = pd.Series(range(10000))

In [2]: f = lambda x: np.sum(x) + 5

# row is unused; row is always an ndarray
In [4]: %timeit s.rolling(10).apply(f, raw=False)
219 ms ± 29.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: %timeit s.rolling(10).apply(f, raw=True)
226 ms ± 26.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

```
# Master
In [1]: s = pd.Series(range(10000))

In [2]: f = lambda x: np.sum(x) + 5

# row passed as a Series
In [4]: %timeit s.rolling(10).apply(f, raw=False)
1.19 s ± 5.81 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# row passed as a ndarray
In [5]: %timeit s.rolling(10).apply(f, raw=True)
32.3 ms ± 366 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```