@mroeschke 2019-09-18
Additionally, `apply` currently supports passing `*args` and `**kwargs` into the function per row. When njitting the passed function from the user, passing kwargs is currently unsupported. https://github.com/numba/numba/issues/2916. And binding these kwargs with `functools.partial` beforehand is unsupported, https://github.com/numba/numba/issues/4587