@mroeschke 2019-09-19
I was able to solve the performance problem in https://github.com/twosigma/pandas/pull/29/commits/aa9644c3b1e5e1cc39a62d4fc47a5c6a7c04bf4a. The biggest issue was compiling the `njit` rolling_apply function every time. If we dynamically create the rolling apply function with the passed argument, cache the function, and call it again, performance beats cython.

```
In [1]: s = pd.Series(range(10000))

# r, a Rolling object, will cache the apply functions
In [2]: r = s.rolling(10)

In [3]: f = lambda x: np.sum(x) + 5

In [4]: %timeit r.apply(f, raw=False)
2.16 ms ± 204 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```