@mroeschke 2019-09-25
Here are the ASV benchmarks:

```
       before           after         ratio
     [68a663be]       [3c49d034]
     <master>         <feature/rolling_apply_numba>
+         289±2ms        350±0.9ms     1.21  rolling.Apply.time_rolling('Series', 1000, 'float', <function sum at 0x105754620>, True)
+         290±7ms          350±1ms     1.20  rolling.Apply.time_rolling('Series', 1000, 'int', <function sum at 0x105754620>, True)
+         297±8ms          353±2ms     1.19  rolling.Apply.time_rolling('DataFrame', 1000, 'int', <function sum at 0x105754620>, True)
-         250±3ms        215±0.6ms     0.86  rolling.Apply.time_rolling('Series', 10, 'float', <function sum at 0x105754620>, True)
-         255±1ms        215±0.8ms     0.84  rolling.Apply.time_rolling('Series', 10, 'int', <function sum at 0x105754620>, True)
-         308±5ms        246±0.9ms     0.80  rolling.Apply.time_rolling('Series', 10, 'float', <function Apply.<lambda> at 0x11b8c6730>, True)
-        310±20ms          241±4ms     0.78  rolling.Apply.time_rolling('DataFrame', 10, 'float', <function Apply.<lambda> at 0x11b8c6730>, True)
-        340±40ms        245±0.6ms     0.72  rolling.Apply.time_rolling('Series', 10, 'int', <function Apply.<lambda> at 0x11b8c6730>, True)
-      13.9±0.06s          380±1ms     0.03  rolling.Apply.time_rolling('Series', 1000, 'int', <function Apply.<lambda> at 0x11b8c6730>, False)
-       14.0±0.1s        380±0.8ms     0.03  rolling.Apply.time_rolling('Series', 1000, 'float', <function Apply.<lambda> at 0x11b8c6730>, False)
-         14.1±0s          380±1ms     0.03  rolling.Apply.time_rolling('DataFrame', 1000, 'float', <function Apply.<lambda> at 0x11b8c6730>, False)
-      13.7±0.01s        349±0.8ms     0.03  rolling.Apply.time_rolling('Series', 1000, 'int', <function sum at 0x105754620>, False)
-      13.9±0.06s        350±0.6ms     0.03  rolling.Apply.time_rolling('Series', 1000, 'float', <function sum at 0x105754620>, False)
-      13.7±0.07s        245±0.8ms     0.02  rolling.Apply.time_rolling('Series', 10, 'float', <function Apply.<lambda> at 0x11b8c6730>, False)
-       14.0±0.1s        246±0.5ms     0.02  rolling.Apply.time_rolling('Series', 10, 'int', <function Apply.<lambda> at 0x11b8c6730>, False)
-      13.5±0.01s        215±0.3ms     0.02  rolling.Apply.time_rolling('Series', 10, 'float', <function sum at 0x105754620>, False)
```

The `raw=False` benchmarks are partially misleading since it means cython is handling pandas objects while numba is always handling numpy arrays (numba cannot operate in nopython mode with pandas objects).