@mroeschke 2019-09-25
Additional notes:

- I had to xfail a couple of tests. To operate rolling.apply in `nopython` mode for max performance, apply cannot accept arbitrary functions.
- the passed function to apply cannot accept *args and **kwargs because it's unsupported in numba.